4819 finish Modal refactoring
